installed black and isort

+++ lambda_fns/delete_order/delete_order.py
@@ -1,7 +1,8 @@
-import boto3
-import os
-import json
 import decimal
+import json
+import os
+
+import boto3
 
 TABLE_NAME = os.environ.get("ORDER_TABLE")
 dynamodb = boto3.resource("dynamodb")
@@ -9,19 +10,12 @@ table = dynamodb.Table(TABLE_NAME)
 
 
 def delete_order(event):
-    order_id = event['arguments']['id']
-    response = table.delete_item(
-        Key={
-            'id': order_id,
-            'user_id': 'demo_user'
-        }
-    )
-    print(f'delete_order {order_id} response: {response}')
+    order_id = event["arguments"]["id"]
+    response = table.delete_item(Key={"id": order_id, "user_id": "demo_user"})
+    print(f"delete_order {order_id} response: {response}")
     return response
 
 
 def handler(event, context):
-
-
     delete_order(event)
-    return event['arguments']['id']
+    return event["arguments"]["id"]
